feat: import a second model into a container via anchored-alignment merge; stream large file writes

+++ src/Amql.Cli/ModelExporter.cs
@@ -117,7 +117,7 @@ public static class ModelExporter
             outDir,
             container.Index.Model,
             payloads.Count,
-            payloads.Sum(p => p.Data.Length),
+            payloads.Sum(p => (long)p.Data.Length),
             notes);
     }
